Bound Memento 1.x releases to pre-1.3
Previously, Memento would store UTC ZonedDateTimes in the records and explicitly depended on TimeZones.jl. In 1.3, it switched to only storing `DateTime` in order to make TimeZones.jl an optional dependency. Since the rest of the Julia ecosystem defaults to `localzone` rather than UTC it made sense to update both packages to follow that same conventions.

https://github.com/invenia/Memento.jl/pull/182#issuecomment-965730317